fix affiliates loading states
