+++ src/views/Affiliates/cards/ProgramStatusCard.tsx
@@ -14,23 +14,11 @@ import { useURLConfigs } from '@/hooks/useURLConfigs';
 import { Button } from '@/components/Button';
 import { Icon, IconName } from '@/components/Icon';
 import { Output, OutputType } from '@/components/Output';
-import { OnboardingTriggerButton } from '@/views/dialogs/OnboardingTriggerButton';
 
 interface IProgramCardProps {
   className?: string;
   isVip: boolean;
 }
-
-const ConnectWallet = () => {
-  const stringGetter = useStringGetter();
-
-  return (
-    <div tw="flex flex-col items-center justify-center gap-y-1 px-4 py-2 text-center">
-      <p>{stringGetter({ key: STRING_KEYS.AFFILIATE_CONNECT_WALLET })}</p>
-      <OnboardingTriggerButton />
-    </div>
-  );
-};
 
 export const ProgramStatusCard = ({ className, isVip = false }: IProgramCardProps) => {
   const stringGetter = useStringGetter();
@@ -76,8 +64,6 @@ export const ProgramStatusCard = ({ className, isVip = false }: IProgramCardProp
 
   return (
     <$Container className={className}>
-      {!dydxAddress && <ConnectWallet />}
-
       {dydxAddress && (
         <div tw="flex flex-col gap-y-1 p-1">
           <$Header tw="flex items-center">
@@ -124,7 +110,7 @@ export const ProgramStatusCard = ({ className, isVip = false }: IProgramCardProp
   );
 };
 
-const $Container = tw.div`h-full rounded-0.625 bg-color-layer-3`;
+const $Container = tw.div`rounded-0.625 bg-color-layer-3`;
 
 const $Header = styled.div`
   font-size: 18px;
